ADM-438 extract response and request params structure to public models folder
