+++ frontend/src/clients/ReportClient.ts
@@ -1,28 +1,6 @@
 import { HttpClient } from '@src/clients/Httpclient'
-import { ReportRes } from '@src/types/reportRes'
-
-export interface ReportReq {
-  metrics: string[]
-  pipeline?: {
-    type: string
-    token: string
-  }
-  board?: {
-    token: string
-    type: string
-    site: string
-    email: string
-    projectKey: string
-    boardId: string
-  }
-  sourceControl?: {
-    type: string
-    token: string
-  }
-  calendarType: string
-  startTime: string | null
-  endTime: string | null
-}
+import { ReportRes } from '@src/models/response/reportRes'
+import { ReportReq } from '@src/models/request/reportReq'
 
 export class ReportClient extends HttpClient {
   reportResponse: ReportRes = {
