+++ frontend/src/models/request/boardReq.ts
@@ -1,0 +1,9 @@
+export interface VerifyBoardReq {
+  token: string
+  type: string
+  site: string
+  projectKey: string
+  startTime: string | number | null
+  endTime: string | number | null
+  boardId: string
+}
