+++ frontend/src/models/request/pipelineReq.ts
@@ -1,0 +1,6 @@
+export interface VerifyPipelineReq {
+  type: string
+  token: string
+  startTime: string | number | null
+  endTime: string | number | null
+}
