+++ frontend/src/models/request/sourceControlReq.ts
@@ -1,0 +1,6 @@
+export interface VerifySourceControlReq {
+  type: string
+  token: string
+  startTime: string | number | null
+  endTime: string | number | null
+}
